# swagger_server/models/create_geo_data_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
import re  # noqa: F401,E501
from swagger_server import util


class CreateGeoDataResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, data_id: str=None):  # noqa: E501
        """CreateGeoDataResponse - a model defined in Swagger

        :param data_id: The data_id of this CreateGeoDataResponse.  # noqa: E501
        :type data_id: str
        """
        self.swagger_types = {
            'data_id': str
        }

        self.attribute_map = {
            'data_id': 'dataId'
        }
        self._data_id = data_id

    @classmethod
    def from_dict(cls, dikt) -> 'CreateGeoDataResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateGeoDataResponse of this CreateGeoDataResponse.  # noqa: E501
        :rtype: CreateGeoDataResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def data_id(self) -> str:
        """Gets the data_id of this CreateGeoDataResponse.


        :return: The data_id of this CreateGeoDataResponse.
        :rtype: str
        """
        return self._data_id

    @data_id.setter
    def data_id(self, data_id: str):
        """Sets the data_id of this CreateGeoDataResponse.


        :param data_id: The data_id of this CreateGeoDataResponse.
        :type data_id: str
        """
        if data_id is None:
            raise ValueError("Invalid value for `data_id`, must not be `None`")  # noqa: E501

        self._data_id = data_id